+++ tests/roots/test-base/target/example_model.py
@@ -26,8 +26,10 @@ class ExampleModel(BaseModel):
 
         """
 
-        if not len(v) < 10:
+        if len(v) >= 10:
             raise ValueError("No more than 10 characters allowed")
+
+        return v
 
     class Config:
         env_prefix = "foo_"