+++ tests/roots/test-base/target/example_setting.py
@@ -28,8 +28,10 @@ class ExampleSettings(BaseSettings):
 
         """
 
-        if not len(v) < 10:
+        if len(v) >= 10:
             raise ValueError("No more than 10 characters allowed")
+
+        return v
 
     class Config:
         env_prefix = "foo_"