+++ changelog.rst
@@ -1,6 +1,33 @@
 Changelog
 =========
 
+v1.3.1 - 2021-XX-XX
+-------------------
+
+This is a minor release providing support for ``root_validator``.
+
+Documentation
+~~~~~~~~~~~~~
+
+- Add example showing representation of asterisk and root validators.
+
+Testing
+~~~~~~~
+
+- Add explicit tests for asterisk and root validators.
+
+Internal
+~~~~~~~~
+
+- Refactor and simplify field validator mapping generation of ``inspection.ModelWrapper``.
+
+Contributors
+~~~~~~~~~~~~
+
+- Thanks to `roguh <https://github.com/roguh>`__ for submitting a feature request
+  for ``root_validators`` `#20 <https://github.com/mansenfranzen/autodoc_pydantic/issues/20>`__.
+
+
 v1.3.0 - 2021-05-10
 -------------------
 
@@ -44,7 +71,7 @@ Added
 ~~~~~
 
 - ``show_versions`` function to show important dependency information which are
- relevant for tracking down bugs as part of the new ``utility`` module.
+  relevant for tracking down bugs as part of the new ``utility`` module.
 
 v1.2.0 - 2021-05-09
 -------------------